♻️ Refactor:  bottomNav를 Navigation으로 수정 Ref: QTVY-79
