--- src/components/common/BottomNav/index.ts
@@ -1,3 +1,0 @@
-// components/common/BottomNav/index.ts
-export * from './BottomNav'
-export * from './types'
