+++ src/components/common/Navigation/Navigation.module.css
@@ -1,3 +1,3 @@
-.bottomNav {
+.navigation {
   @apply fixed bottom-0 left-0 right-0 z-50 flex h-24 min-h-24 items-center justify-around bg-primary px-4 pb-safe-bottom sm:h-24 md:h-24 lg:h-28;
 }
