+++ src/components/common/Navigation/index.ts
@@ -1,0 +1,3 @@
+// index.ts
+export * from './Navigation';
+export * from './types';
